@@ -18,14 +18,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.15.0
-  hooks_riverpod: ^0.12.1
-  flutter_riverpod: ^0.12.1
+  flutter_hooks: ^0.18.0
+  hooks_riverpod: ^1.0.0-dev.6
+  flutter_riverpod: ^1.0.0-dev.6
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
