@@ -18,35 +18,37 @@ class BooleanNotifier extends StateNotifier<bool> {
 
 final booleanProvider = StateNotifierProvider((_) => BooleanNotifier(false));
 
-final booleanState = Provider((ref) => ref.watch(booleanProvider.state));
+final booleanState = Provider((ref) => ref.watch(booleanProvider));
 
-class MemoryExample extends StatelessWidget {
+class MemoryExample extends HookConsumerWidget {
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
+    final notifier = ref.watch(booleanProvider.notifier);
     return MaterialApp(
       theme: ThemeData(
         primarySwatch: Colors.green,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
       home: Scaffold(
-          appBar: AppBar(
-            title: Text("Riverpod Animation Example"),
-          ),
-          body: Column(
-            children: [
-              AnimatedWidget(),
-              RaisedButton(
-                  child: Text("Rotate"),
-                  onPressed: () {
-                    context.read(booleanProvider).toggle();
-                  }),
-            ],
-          )),
+        appBar: AppBar(
+          title: Text("Riverpod Animation Example"),
+        ),
+        body: Column(
+          children: [
+            AnimatedWidget(),
+            RaisedButton(
+                child: Text("Rotate"),
+                onPressed: () {
+                  notifier.toggle();
+                }),
+          ],
+        ),
+      ),
     );
   }
 }
 
-class AnimatedWidget extends HookWidget {
+class AnimatedWidget extends HookConsumerWidget {
   final Duration duration = const Duration(seconds: 1);
 
   Color getColor(bool isUp) {
@@ -64,11 +66,11 @@ class AnimatedWidget extends HookWidget {
   }
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     final _controller =
         useAnimationController(duration: duration, initialValue: 1);
 
-    var _isUp = useProvider(booleanState);
+    var _isUp = ref.watch(booleanState);
     final isUp = useState(_isUp);
 
     useValueChanged(_isUp, (_, __) async {
@@ -78,26 +80,27 @@ class AnimatedWidget extends HookWidget {
     });
 
     return AnimatedBuilder(
-        animation: _controller,
-        builder: (context, child) {
-          return Transform(
-            transform: Matrix4.rotationX((1 - _controller.value) * pi / 2),
+      animation: _controller,
+      builder: (context, child) {
+        return Transform(
+          transform: Matrix4.rotationX((1 - _controller.value) * pi / 2),
+          alignment: Alignment.center,
+          child: Container(
+            height: 100,
+            margin: EdgeInsets.all(20),
+            padding: EdgeInsets.all(12),
             alignment: Alignment.center,
-            child: Container(
-              height: 100,
-              margin: EdgeInsets.all(20),
-              padding: EdgeInsets.all(12),
-              alignment: Alignment.center,
-              decoration: BoxDecoration(
-                  color: getColor(isUp.value),
-                  border: Border.all(color: Colors.grey)),
-              child: Icon(
-                getIcon(isUp.value),
-                size: 40,
-                color: Colors.white70,
-              ),
+            decoration: BoxDecoration(
+                color: getColor(isUp.value),
+                border: Border.all(color: Colors.grey)),
+            child: Icon(
+              getIcon(isUp.value),
+              size: 40,
+              color: Colors.white70,
             ),
-          );
-        });
+          ),
+        );
+      },
+    );
   }
 }
